Class "Model" renamed to "Battleground", created interface Model
