+++ src/main/java/com/lilangel/model/Model.java
@@ -2,41 +2,10 @@ package com.lilangel.model;
 
 import com.lilangel.presenter.ModelListener;
 
-import java.util.ArrayList;
-
 /**
- * Class that describes the model of battleground and the robots
+ * Inteface for the model
  */
-public class Model {
-    ModelListener presenter;
-    private double[][] battleGround;
-    private final ArrayList<Robot> robots;
-    private final int robotsCount = 20; //пока 20, потом посмотрим, может можно больше
-
-    public Model() {
-        this.battleGround = new double[150][150];
-        this.robots = new ArrayList<>(robotsCount);
-        for(int i =0; i<robotsCount;i++)
-            robots.add(new Robot());
-    }
-
-    public void setPresenter(ModelListener presenter) {
-        this.presenter = presenter;
-    }
-
-    public void startFight() {
-
-    }
-
-    public ModelState getState() {
-        return new ModelState(robots);
-    }
-
-    public double[][] getBattleGround() {
-        return battleGround;
-    }
-
-    private void setBattleGround(double[][] battleGround) {
-        this.battleGround = battleGround;
-    }
+public interface Model {
+    void notifyPresenter();
+    void setPresenter(ModelListener presenter);
 }
